+++ src/main/java/com/training/model/dao/mappers/DiscountMapper.java
@@ -32,20 +32,19 @@ public class DiscountMapper  extends AbstractMapper<Discount>{
             log.debug("destinationStreetId is empty");
         }
         try {
-            // TODO: 4/20/19 change datatype in table from int to long
-            discount.setMinimalBill((long)(int) set.getObject("minimalBill"));
+            discount.setMinimalBill((long) set.getObject("minimalBill"));
             log.debug("minimalBill:" + discount.getMinimalBill());
         } catch (NullPointerException e){
             log.debug("minimalBill is empty");
         }
         try {
-            discount.setMinimalThreshold((long)(int) set.getObject("minimalThreshold"));
+            discount.setMinimalThreshold((long) set.getObject("minimalThreshold"));
             log.debug("minimalThreshold:" + discount.getMinimalThreshold());
         } catch (NullPointerException e){
             log.debug("minimalThreshold is empty");
         }
         try {
-            discount.setDiscount((long)(int) set.getObject("discount"));
+            discount.setDiscount((long) set.getObject("discount"));
             log.debug("discount:" + discount.getDiscount());
         } catch (NullPointerException e){
             log.debug("discount is empty");
